+++ flake.nix
@@ -1,10 +1,15 @@
 {
   inputs = {
     flake-utils.url = "github:numtide/flake-utils";
-    nixpkgs.url = "github:nixos/nixpkgs/nixos-23.11";
+    nixpkgs.url = "github:nixos/nixpkgs/nixos-24.11";
   };
 
-  outputs = { self, flake-utils, nixpkgs }:
+  outputs =
+    {
+      self,
+      flake-utils,
+      nixpkgs,
+    }:
     flake-utils.lib.simpleFlake {
       inherit self nixpkgs;
       name = "rollbar-haskell";